tigu_backend_fastapi/pyproject.toml
[tool.poetry]
name = "tigu-backend-fastapi"
version = "0.1.0"
description = "FastAPI backend for Tigu platform"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.100.0" # Or your preferred version
uvicorn = {extras = ["standard"], version = "^0.23.2"}
sqlalchemy = "^2.0.0"
pydantic = {extras = ["email"], version = "^2.0.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
alembic = "^1.11.1"
psycopg2-binary = "^2.9.6" # Or asyncpg for async
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
httpx = "^0.24.1" # For testing FastAPI async clients

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
